파일 구조 개선: 컨트롤러 분리 및 메인 선택 화면 추가
✨ 주요 개선사항:
- ViewController.swift를 메인 선택 화면으로 개편
- TableViewController.swift 별도 파일로 분리 (다양한 스타일 예제)
- LabelViewController.swift 별도 파일로 분리 (애니메이션 예제)
- SceneDelegate 네비게이션 구조 업데이트

🏗️ 새로운 구조:
- 메인 화면에서 두 가지 예제 선택 가능
- 각 예제별로 독립된 뷰 컨트롤러
- 명확한 책임 분리와 코드 모듈화

📱 사용자 경험 개선:
- 직관적인 예제 선택 인터페이스
- Large Title 네비게이션으로 현대적 디자인
- iOS 16+ 기능 체크 및 적절한 안내

🔧 기술적 개선:
- XcodeBuildMCP 통합 빌드 테스트 완료
- Xcode 프로젝트 파일 수동 업데이트
- 모든 참조 및 의존성 관계 정리
